requirements.txt
basicsr
pillow
torch==2.0.1
torchvision==0.15.2
pandas 
tqdm
pytorch-lightning==1.9.5
timm==0.9.2
h5py
PyYAML
albumentations
pickle-mixin
parmap
opencv-python
openslide-python
matplotlib
huggingface_hub==0.22.2
chardet
transformers
jupyter
wandb